+++ tests/testthat/test-verb-copy-to.R
@@ -35,6 +35,16 @@ test_that("only overwrite existing table if explicitly requested", {
   expect_silent(copy_to(con, data.frame(x = 1), name = "df", overwrite = TRUE))
 })
 
+test_that("overwrite flag works when copying *within* db sources", {
+  con <- local_sqlite_connection()
+  df <- copy_to(con, tibble(x = 1), "df", temporary = FALSE)
+  copy_to(con, df, "df2", temporary = FALSE)
+
+  df2 <- tibble(x = 2)
+  out <- copy_to(con, df2, name = "df2", temporary = FALSE, overwrite = TRUE)
+  expect_equal(collect(out), df2)
+})
+
 test_that("can create a new table in non-default schema", {
   con <- local_sqlite_con_with_aux()